Fixed reminder scheduling bugs after task was either completed, uncompleted, deleted or undeleted in different client. Those actions caused task to be updated and furthermore active reminder to be marked removed or not removed and then the task saved. Last http request, basically analytics data, blocked the http request queue hence blocking subsequent saveTask calls. Fixed by releasing http request queue lock immediately when the analytics request is executed. Prevent sending analytics request twice as well.

+++ frontend/core/src/main/webapp/static/app/base/http/HttpClientService.js
@@ -63,20 +63,28 @@
 
 
   function executeLastRequest(lastRequest){
-    $.ajax({
-      type: lastRequest.content.method.toUpperCase(),
-      url: lastRequest.content.url,
-      data: JSON.stringify(lastRequest.content.data),
-      contentType: 'application/json',
-      dataType: 'json',
-      success: function() {
-        // Then remove the request and release lock
-        HttpRequestQueueService.remove(lastRequest);
-      },
-      error: function() {
-        HttpRequestQueueService.releaseLock();
-      }
-    });
+    // As last request is insignificant, lock needs to be released before executing to prevent
+    // problems with subsequent calls failing
+    HttpRequestQueueService.releaseLock();
+
+    if (!lastRequest.executing){
+      lastRequest.executing = true;
+      $.ajax({
+        type: lastRequest.content.method.toUpperCase(),
+        url: lastRequest.content.url,
+        data: JSON.stringify(lastRequest.content.data),
+        contentType: 'application/json',
+        dataType: 'json',
+        success: function() {
+          // Then remove the request (lock was released before execution)
+          HttpRequestQueueService.remove(lastRequest);
+          delete lastRequest.executing;
+        },
+        error: function() {
+          delete lastRequest.executing;
+        }
+      });
+    }
   }
 
   var retryingExecution = false;
